change naming and separate subscribe function
