+++ database/userevent.py
@@ -1,9 +1,8 @@
 from sqlalchemy import Boolean, Column, Integer, String
-
 from infrastructure.mysql import Base
 
 
-class Subscribe(Base):
+class UserEvent(Base):
     __tablename__ = 'users_events'
 
     id = Column(Integer, primary_key=True)
